Update README.md
Aim- 
_________
The handwritten numeral image recognition problem becomes one of the most famous areas of research in the field of machine learning and computer vision. This project seeks to provide a solution to the digitization of the given input image of a handwritten numeral digit. The main aim of this project to convert real-time input images of handwritten digits into the text as well as an audio files. 

Technologies Undertaken- 
______________________________
1. Machine Learning (Supervised Approach)
2. Digital Image Processing 
3. Python 3.x- Jupyter Notebook 
4. MNIST Dataset (0-9 digits) 

Methodology- 
_________________
1. Image Acquinsion- Grab the input image 
2. Image Processing- Pretreatment of User Input 
3. Feature Extraction 
4. Model Training 
5. Prediction 
6. Text and Audio Files corresponding to the predicted output (optional step) 

GUI- 
________
 The graphical User Interface for this project allows ease to use the system even for any non-technical person. Moreover, it has two color themes- 
1. Dark Mode 
2. Bright Mode 

Dataset-
___________
The MNIST database consisting of offline handwritten digits ranging from 0-9, downloaded from Kaggle. This dataset contains Train.csv and Test.csv datafiles. 

The Train.csv file is used as a training dataset and it contains 42000 greyscale images of hand-drawn digits from 0-9. Each digit is a gray-level fixed-size image with a size of 28 x 28 (or 784 pixels) in total as the features. For this model, 75% training set, and 25% testing set. 

Each pixel has a single pixel value associated with it and these pixel values are integers ranging between 0-255 inclusive. Training Dataset has 785 columns & 42001 rows. The 1st column holds labels, which is the digit that was drawn by the user. The rest of the columns contains the pixel values of the associated images. The test set is the same as the training set, except that it doesn’t contain the label column. 

ML Model used-
___________________
 Support Vector Machine using RBF Kernel- In machine learning, the radial basis function kernel, or RBF kernel, is a popular kernel function used in various kernelized learning algorithms. In particular, it is commonly used in support vector machine classification. RBF is giving for another class, what factors they depend upon and the information we can get from it. Kernels in SVM classification refer to the function that is responsible for defining the decision boundaries between the classes. Apart from the classic linear kernel which assumes that the different classes are separated by a straight line, an RBF (radial basis function) kernel is used when the boundaries are hypothesized to be curve-shaped. RBF kernel uses two main parameters, gamma, and C that are related to: 1. the decision region (how spread the region is), and 2. the penalty for misclassifying a data point In machine learning, the radial basis function kernel, or RBF kernel, is a popular kernel function used in various kernelized learning algorithms. In particular, it is commonly used in support vector machine classification. RBF is giving for other classes, what factors they depend upon and the information we can get from it. Kernels in SVM classification refer to the function that is responsible for defining the decision boundaries between the classes. Apart from the classic linear kernel which assumes that the different classes are separated by a straight line, an RBF (radial basis function) kernel is used when the boundaries are hypothesized to be curve-shaped. RBF kernel uses two main parameters, gamma, and C that are related to: 
1. the decision region (how spread the region is), 
2. the penalty for misclassifying a data point 

Model Accuracy- 
____________________
98% Accurate predicions are made successfully, 

Future scope & Extension-
 ____________________________
Document Analyzer 
Vehicle Number Plate Character Recognizer 
Postal Address Recognizer 
We can add Multi Language Feature 
It can help blind and illiterate people by adding a speech module